[#140488815] Improve forward compatibility (#23)
* [#140488815] Improve forward compatibility

In order to make the code forward compatible, some changes had to be
made. The following requirements had to be met:

* Ignore extra JSON fields
* Handle missing JSON fields (add null checks)
* Ignore extra HTTP end points
* Handle missing HTTP end points

These JSON field requirements apply to both the renos and the client
side, although this commit only applies to the example code (client).

In order to achieve this, null checks were added to all the fields
that are to be received in JSON format over the HTTP/WS communication.
Additionally, error messages were added to handle missing end points.

Note: Sending an empty string instead of JSON will cause a
NullPointerException on the renos side. It would be a good idea to
handle this in a more robust way. My proposal would be to ignore empty
strings and log (debug) that an empty string was received and ignored.

Other than forward compatibility, some convenient fixes were done:

* Replace "throws Exception" with specific exceptions
* Add custom exceptions
* Check JSON parsing success at a central location
* Check user input parsing succes at a central location

Concerns: #140488815